Improved code efficiency
Cut down the creation of a csv file to one line compared to several through the use of pipes (|)

NOTE: Currently when creating a json file is created, the csv file it is converted from gets removed. This could cause problems if wanting to have both csv and json files respectively. Will look to amend.